qi.eventloop: Catches and logs exceptions when spawning threads fails. #43512
When creating a thread throws, we log an error and skip the creation
for the "grace period". This is to avoid the whole application to crash
in the situation where the system doesn't allow acquiring the resources
to create the thread. In this case, the application can still run with
the already allocated threads.
If the issue is triggered by memory exhaustion, the application will
probably crash at the next allocation anyway.

Change-Id: Ib15643e51eda6afb2a3044709a95406effb12008
Reviewed-on: http://gerrit.aldebaran.lan/94421
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: vincent.palancher <vincent.palancher@external.softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jlamotte <jlamotte@aldebaran.com>